Significantly reduced the file-size of graphical assets to reduce overhead when initializing.
Some of these files have been reduced by 92%! The fluffy default avatar had no business being nearly half a meg, nor be larger than we resize actual character avatars to when the card is first assembled... The nature of the use of the fluffy placeholder avatar, as well as the "This image is 99% white, red, and black pixels." allowed the bit-depth to be reduced to 4-bits per pixel (Was 32bpp) and being only a placeholder/fallback avatar image, after much consideration it was dropped to 1/2 of maximum avatar dimensions as well. fluffy.png is now the appropriate 4x6 WxH ratio that card avatars are cropped to, and at 200x300 pixels (Half of the maximum 400x600 that character avatars are resized to.) at 4 bits each, she went from 418KiB to all of 31.5KiB and loads in a tenth of the time now.

The rest of the images have NOT had their dimensions altered, merely changed from a depth of 32bits per pixel, to a much more prudent 8-bits per pixel and examining the "difference" map between before and after, there has been no visible detriment to the perceived quality or sharpness. (Many of them are derived from limited-palette digital paintings, and as such 32bpp is a very inefficient use of both space and causes contention surges for those with lower-bandwidth connections (4G/LTE mobile and stationary ADSL connections easily saturated during the initialization surge when the browser attempts to retrieve all graphical assets, especially for those with 50+ 400x600 32bpp card avatars all trying to download simultaneously. Processed 7 additional assets in public/img/ to minimize file-size with no discernible loss of visual quality to hasten loading times and reduce the footprint of the browser-memory overhead, some of these images are not even called in the code, but because the initialization process iterates through the entire contents of the directory, they still wind up being pre-cached into the browser.

Processed several images in the ./readme/ folder. These do not get pre-loaded into the browser session, but they do consume bandwidth every time someone pulls the repo via the zip file, or clones a fresh copy... Waste-not, want-not.

<3 ya Humi.

--FunkEngine