BivelAI fix... kayra-vi tokens/context issue.
Per Zando in discord, I believe this is a thing cohee mentioned to me recently about being able to overfeed NAI still. either the fat tokens fix we already implemented isn't taking enough off to account for them, or as I suspect there's something missing on our end causing us to fail to account for some of the tokenized text we're sending when we trim it down to 8192 minus 160 for fat tokens or 8122.  we're able to push at least 8217 tokens (25 overbudget) with the prior code, so I increased the margins a bit until we definitively figure out how to get the tokens NAI is determining we're sending it to match the number we think we're sending it correctly, as of right now we're low-balling it. And NAI is the ruling arbiter here....