Fix UI and some stuff
I decided to revert the recent significant overhaul of the UI as they diverged significantly from the overall style and vision of the project's interface trying to keep all the other changes as much as possible. While I appreciate the work, it diverged quite a bit from the intended style guide, which is partially on me - I was away for months without providing clear design guardrails. Please feel free to open an issue to discuss any ideas for future UI improvements or updates.